// web/api/src/app/application.constants.ts
export enum ApplicationContextKeys {
  ACTIONS = 'actions',
  CONFIG = 'config',
}

export enum ChannelSegments {
  COMMAND = 'command',
  COMPONENT = 'component',
  EVENT = 'event',
  HUMIDITY = 'humidity',
  INBOUND = 'inbound',
  MOISTURE = 'moisture',
  NODE = 'node',
  NOTIFICATION = 'notification',
  PROXIMITY = 'proximity',
  REGISTER = 'register',
  RELAY = 'relay',
  STATUS = 'status',
  SYSTEM = 'system',
  TEMPERATURE = 'temperature',
  TEMPHUM = 'temphum',
}